verify settings is here! (sorta)
So it does indeed check if the setting exists, albeit in a funky ass way. The main reason its sucks is cause for keybinds, pretty much any isAlpha() char would work, and that makes it a lil bit funky. if youve got any ideas to improve on t hat other then the horrid if then, lemme know. Anywho, imma finally begin trying the sfml thingie on to a new branch!