Update CMakeLists.txt
Updated CMakeLists.txt to correct some messages and conditions.

Fatal error on line 39 would be thrown for any out-of-source builds not in the /build/ directory, and claim that an in-source build threw the error.

Condition on line 121 was checking if it was running on a 32 bit system, when it should have checked if it was not running on a 64 bit system.