Updated main.cpp
Updated main.cpp to make test program more visual.

Tried running the program on Ubuntu and was greeted with a black screen. Thought the program had failed before checking the code. This commit adds a red circle to the screen.

While it may not fit will with the purpose of a prerelease branch, we're early enough in development that it shouldn't matter.