Finished Mac tests of steamwork library
Fixed issue #11 ( as far we can tell ). Instead of moving dynamic libraries to build/bin, this creates a post_build script to move the dynamic libraries to wherever the executable is.

For buildingReleases.txt (not in this branch) we'll no longer have to copy the dynamic libraries from build/bin to the release/ folder.

Also completed to-do list item of testing steamworks install on mac.